+++ src/Compilers/Test/Core/Compilation/RuntimeUtilities.cs
@@ -2,10 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#nullable disable
-
 using System;
 using System.Collections.Immutable;
+using System.Globalization;
 using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
@@ -18,6 +17,8 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
     /// </summary>
     public static partial class RuntimeUtilities
     {
+        private static readonly object s_outputGuard = new();
+
         internal static bool IsDesktopRuntime =>
 #if NET472
             true;
@@ -30,7 +31,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
 
         private static int? CoreClrRuntimeVersion { get; } = IsDesktopRuntime
             ? null
-            : typeof(object).Assembly.GetName().Version.Major;
+            : typeof(object).Assembly.GetName()!.Version!.Major;
 
         internal static bool IsCoreClr6Runtime
             => IsCoreClrRuntime && RuntimeInformation.FrameworkDescription.StartsWith(".NET 6.", StringComparison.Ordinal);
@@ -41,7 +42,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
         internal static bool IsCoreClr9OrHigherRuntime
             => CoreClrRuntimeVersion is { } v && v >= 9;
 
-        internal static BuildPaths CreateBuildPaths(string workingDirectory, string sdkDirectory = null, string tempDirectory = null)
+        internal static BuildPaths CreateBuildPaths(string workingDirectory, string? sdkDirectory = null, string? tempDirectory = null)
         {
             tempDirectory ??= Path.GetTempPath();
 #if NET472
@@ -70,23 +71,40 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
 #endif
         }
 
-        internal static (string Output, string ErrorOutput) CaptureOutput(Action action)
-        {
-#if NET472
-            return Roslyn.Test.Utilities.Desktop.DesktopRuntimeEnvironment.CaptureOutput(action);
-#elif NETCOREAPP
-            return Roslyn.Test.Utilities.CoreClr.CoreCLRRuntimeEnvironment.CaptureOutput(action);
-#else
-#error Unsupported configuration
-#endif
-        }
-
         /// <summary>
         /// Get the location of the assembly that contains this type
         /// </summary>
         internal static string GetAssemblyLocation(Type type)
         {
             return type.GetTypeInfo().Assembly.Location;
+        }
+
+        public static (string Output, string ErrorOutput) CaptureOutput(Action action, IFormatProvider? formatProvider = null)
+        {
+            formatProvider ??= CultureInfo.InvariantCulture;
+            lock (s_outputGuard)
+            {
+                var savedConsoleOut = Console.Out;
+                var savedConsoleError = Console.Error;
+
+                using var outputWriter = new StringWriter(formatProvider);
+                using var errorWriter = new StringWriter(formatProvider);
+                try
+                {
+                    Console.SetOut(outputWriter);
+                    Console.SetError(errorWriter);
+                    action();
+                }
+                finally
+                {
+                    Console.SetOut(savedConsoleOut);
+                    Console.SetError(savedConsoleError);
+                }
+
+                var output = outputWriter.ToString();
+                var errorOutput = errorWriter.ToString();
+                return (output, errorOutput);
+            }
         }
     }
 }
